+++ src/utils/index.js
@@ -1,4 +1,3 @@
-export batchHandler from "./batch-handler";
 export exitHandler from "./exit-handler";
 export notifHandler from "./notif-handler";
 export oAuthHandler from "./oauth-handler";
@@ -10,7 +9,6 @@ export staticRouter from "./static-router";
 export tokenMiddleware from "./token-middleware";
 export requireHullMiddleware from "./require-hull-middleware";
 export responseMiddleware from "./response-middleware";
-export serviceMiddleware from "./service-middleware";
 export notifMiddleware from "./notif-middleware";
 export segmentsMiddleware from "./segments-middleware";
 export helpersMiddleware from "./helpers-middleware";